First project is not finished
I couldnt get around on finishing the first project.
It's an ongoing task for me to do.

I need to raise a flag on this one.
that and i could wrap my head around the whole css thing. a little too complicated and i need to focus on it for a long time to completely understand it